Add files via upload
Configuring dynamic IP EIGRP (Enhanced Interior Gateway Routing Protocol) 
routing with ISP network. This will be done in a LAB of connecting two offices 
together, head office and branch office, the purpose is for both to be able to 
communicate or share date or information through network protocols or let just 
say to route traffics between the branches. However, if it has an organization that 
has more than one branch, that will also be possible. 
 The LAB will be done by placing 2811 Cisco router and 2960 Cisco switch in each 
office. The routers will serve as a DHCP server and will connect both office together 
and each router in the office will be connected to the switch to provide network to 
all LAN devices from the router. The project will be done using Cisco Packet Tracer. 
Please have in mind that this is a practical Lab, in real world to connect the 2 
branches together the organization will have to lay a fiber optic cable or use a fixed 
wireless internet (fixed wireless transmit signal through the air from a nearby tower 
which means they provide a cable free connection.